backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
django-rest-swagger==2.2.0
django-versatileimagefield==2.1
django-filter==2.4.0
mysqlclient==2.0.3
phonenumbers==8.12.24
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
twilio==6.61.0
typing-extensions==3.10.0.0
